moved marketing cookie logic into the app, for all pages to consume
